+++ adapters/lora-vertx/src/main/java/org/eclipse/hono/adapter/lora/providers/ThingsNetworkProvider.java
@@ -16,6 +16,7 @@ package org.eclipse.hono.adapter.lora.providers;
 import java.util.Base64;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -63,8 +64,8 @@ public class ThingsNetworkProvider extends JsonBasedLoraProvider {
     }
 
     @Override
-    public String pathPrefix() {
-        return "/ttn";
+    public Set<String> pathPrefixes() {
+        return Set.of("/ttn");
     }
 
     /**